Add files via upload
number_guess - user tries to guess a number and is needed to match it with the number system generated randon number within a range. guess needs to be evaluated for closeness to the random num generated by system. loop until match. if matched display message and exit. (Used random.randint)

Scramble_word - from a list of strings, scramble and display a string. loop until user doesnt input 'Y' to guess or if list if exhausted.